refactor(file): move file-metadata to util
